# packages/contracts/openapi.json
{
  "components": {
    "parameters": {
      "TeachlyRequestId": {
        "description": "Optional caller correlation id. Echoed in the response; generated when omitted.",
        "in": "header",
        "name": "x-request-id",
        "required": false,
        "schema": {
          "maxLength": 200,
          "type": "string"
        }
      }
    },
    "responses": {
...
      "TeachlyTooManyRequests": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorDto"
            }
          }
        },
        "description": "Quota exceeded. Retry after the current rate-limit window."
      }
    },
    "schemas": {
...
      "ApiErrorDto": {
        "properties": {
          "code": {
            "example": "BAD_REQUEST",
            "type": "string"
          },
          "details": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "message": {
            "example": "Validation failed",
            "type": "string"
          },
          "requestId": {
            "example": "phase-2-contract-test",
            "type": "string"
          },
          "statusCode": {
            "example": 400,
            "type": "number"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "requestId"
        ],
        "type": "object"
      },
...
      "ExternalResultObservationDto": {
        "properties": {
          "externalLearnerId": {
            "type": "string"
          },
          "externalTaskId": {
            "type": "string"
          },
          "idempotencyKey": {
            "type": "string"
          },
          "observedAt": {
            "type": "string"
          },
          "outcome": {
            "enum": [
              "correct",
              "incorrect",
              "invalid"
            ],
            "type": "string"
          },
          "score": {
            "type": "number"
          },
          "sourceMetadata": {
            "type": "object"
          },
          "taskSourceId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "externalLearnerId",
          "taskSourceId",
          "externalTaskId",
          "idempotencyKey",
          "observedAt"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "metrics-token": {
        "bearerFormat": "Metrics token",
        "description": "Dedicated token for Prometheus-compatible metrics scraping.",
        "scheme": "bearer",
        "type": "http"
      },
      "workspace-api-key": {
        "bearerFormat": "Teachly workspace API key",
        "description": "Workspace-scoped API key. Never expose it in browser code.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Versioned integration API for the Teachly learning platform",
    "title": "Teachly API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/assessment/results/external": {
      "post": {
        "description": "Auth: workspace API key with scope assessment:write (integration authentication is required). Records an immutable observation and hands off to Learning; replaying the same idempotencyKey returns idempotentReplay=true.",
        "operationId": "TaskBankController_observeExternalResult",
        "parameters": [
          {
            "$ref": "#/components/parameters/TeachlyRequestId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalResultObservationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalResultObservationResponseDto"
                }
              }
            },
            "description": "Recorded external result observation."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Malformed request, validation failure, or invalid UUID/path parameter."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing, invalid, or revoked workspace API key, or no authenticated user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing role, wrong workspace, or missing required assessment scope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "External learner, imported task, or published task version not found."
          },
          "429": {
            "$ref": "#/components/responses/TeachlyTooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/TeachlyInternalError"
          }
        },
        "security": [
          {
            "workspace-api-key": []
          }
        ],
        "summary": "Record an external assessment result",
        "tags": [
          "assessment-task-bank"
        ]
      }
    },
    "/v1/assessment/task-drafts/{draftId}": {
      "patch": {
        "description": "Auth: workspace API key with scope assessment:write, or a user session with a write-capable role. Patching a published version creates a new draft version instead of modifying it.",
        "operationId": "TaskBankController_updateDraft",
        "parameters": [
          {
            "description": "Task version id (draft or published).",
            "in": "path",
            "name": "draftId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/TeachlyRequestId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDraftDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskVersionResponseDto"
                }
              }
            },
            "description": "Updated or newly created draft task version."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Malformed request, validation failure, or invalid UUID/path parameter."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing, invalid, or revoked workspace API key, or no authenticated user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing role, wrong workspace, or missing required assessment scope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Task draft not found."
          },
          "429": {
            "$ref": "#/components/responses/TeachlyTooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/TeachlyInternalError"
          }
        },
        "security": [
          {
            "workspace-api-key": []
          }
        ],
        "summary": "Update a draft task version",
        "tags": [
          "assessment-task-bank"
        ]
      }
    },
    "/v1/assessment/task-drafts/{draftId}/publish": {
      "post": {
        "description": "Auth: workspace API key with scope assessment:manage, or a user session with a manage-capable role. Only draft versions can be published; published versions become immutable.",
        "operationId": "TaskBankController_publish",
        "parameters": [
          {
            "description": "Draft task version id.",
            "in": "path",
            "name": "draftId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/TeachlyRequestId"
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskVersionResponseDto"
                }
              }
            },
            "description": "Published task version."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Malformed request, validation failure, or invalid UUID/path parameter."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing, invalid, or revoked workspace API key, or no authenticated user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing role, wrong workspace, or missing required assessment scope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Task draft not found."
          },
          "429": {
            "$ref": "#/components/responses/TeachlyTooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/TeachlyInternalError"
          }
        },
        "security": [
          {
            "workspace-api-key": []
          }
        ],
        "summary": "Publish a draft task version",
        "tags": [
          "assessment-task-bank"
        ]
      }
    },
    "/v1/assessment/task-sources": {
      "get": {
        "description": "Auth: workspace API key with scope assessment:read, or an authenticated user session.",
        "operationId": "TaskBankController_listSources",
        "parameters": [
          {
            "in": "query",
            "name": "workspaceId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "$ref": "#/components/parameters/TeachlyRequestId"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskSourceResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Task sources of the resolved workspace."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Malformed request, validation failure, or invalid UUID/path parameter."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing, invalid, or revoked workspace API key, or no authenticated user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing role, wrong workspace, or missing required assessment scope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Requested workspace does not exist."
          },
          "429": {
            "$ref": "#/components/responses/TeachlyTooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/TeachlyInternalError"
          }
        },
        "security": [
          {
            "workspace-api-key": []
          }
        ],
        "summary": "List task sources",
        "tags": [
          "assessment-task-bank"
        ]
      },
      "post": {
        "description": "Auth: workspace API key with scope assessment:manage, or a user session with organization_admin, workspace_admin, or content_editor role.",
        "operationId": "TaskBankController_createSource",
        "parameters": [
          {
            "$ref": "#/components/parameters/TeachlyRequestId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskSourceDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSourceResponseDto"
                }
              }
            },
            "description": "Created task source."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Malformed request, validation failure, or invalid UUID/path parameter."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing, invalid, or revoked workspace API key, or no authenticated user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing role, wrong workspace, or missing required assessment scope."
          },
          "429": {
            "$ref": "#/components/responses/TeachlyTooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/TeachlyInternalError"
          }
        },
        "security": [
          {
            "workspace-api-key": []
          }
        ],
        "summary": "Create a task source",
        "tags": [
          "assessment-task-bank"
        ]
      }
    },
    "/v1/assessment/task-sources/{id}": {
      "get": {
        "description": "Auth: workspace API key with scope assessment:read, or an authenticated user session.",
        "operationId": "TaskBankController_getSource",
        "parameters": [
          {
            "description": "Task source id.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/TeachlyRequestId"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSourceResponseDto"
                }
              }
            },
            "description": "Task source."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Malformed request, validation failure, or invalid UUID/path parameter."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing, invalid, or revoked workspace API key, or no authenticated user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing role, wrong workspace, or missing required assessment scope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Task source not found in this tenant."
          },
          "429": {
            "$ref": "#/components/responses/TeachlyTooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/TeachlyInternalError"
          }
        },
        "security": [
          {
            "workspace-api-key": []
          }
        ],
        "summary": "Get a task source",
        "tags": [
          "assessment-task-bank"
        ]
      }
    },
    "/v1/assessment/task-sources/{id}/import": {
      "post": {
        "description": "Auth: workspace API key with scope assessment:write, or a user session with a write-capable role. Replaying the same idempotencyKey returns the previously imported version with idempotentReplay=true.",
        "operationId": "TaskBankController_import",
        "parameters": [
          {
            "description": "Task source id, must be active.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/TeachlyRequestId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportTaskDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportedTaskVersionResponseDto"
                }
              }
            },
            "description": "Imported (or replayed) task version."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Malformed request, validation failure, or invalid UUID/path parameter."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing, invalid, or revoked workspace API key, or no authenticated user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Missing role, wrong workspace, or missing required assessment scope."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "Task source not found in this tenant."
          },
          "429": {
            "$ref": "#/components/responses/TeachlyTooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/TeachlyInternalError"
          }
        },
        "security": [
          {
            "workspace-api-key": []
          }
        ],
        "summary": "Import a task snapshot into a task source",
        "tags": [
          "assessment-task-bank"
        ]
      }
    },
...
  },
  "servers": [],
...
}